configurações locais

+++ aula-01/pix-consumidor/src/main/java/com/alura/pix/repository/KeyRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface KeyRepository extends JpaRepository<Key, Integer> {
+public interface KeyRepository extends JpaRepository<Key, Long> {
     Key findByChave(String key);
 }
